And break out title and description as a separate trait
